mimeo #35 - extract mimeo cli to a package
