+++ .isort.cfg
@@ -1,2 +1,2 @@
 [settings]
-skip_glob=src/mimeo/generators/__init__.py,src/mimeo/consumers/__init__.py,src/mimeo/context/__init__.py
+skip_glob=src/mimeo/cli/__init__.py,src/mimeo/consumers/__init__.py,src/mimeo/context/__init__.py,src/mimeo/generators/__init__.py
